requirements.txt
torch>=2.0
torchvision>=0.15
timm>=1.0
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
scanpy>=1.9
anndata>=0.10
cell2location>=0.1.4
pybiomart>=0.2.0
matplotlib>=3.7
tensorboardX>=2.6
tqdm>=4.60
Pillow>=9.0
